Create Practical4.py
Practical Title : A program that calculates the average of a list of
numbers.
Practical Description: The objective of this practical is to
develop a Python program that calculates the average of a list of
numbers. This exercise will help students understand basic
programming concepts loops, input/output operations, list
datatype.